+++ sentry-spring-boot-starter/api/sentry-spring-boot-starter.api
@@ -44,12 +44,12 @@ public final class io/sentry/spring/boot/SentryTracingCondition : org/springfram
 	public fun <init> ()V
 }
 
-public class io/sentry/spring/boot/datasource/dsproxy/SentryDsProxyAutoConfiguration {
+public class io/sentry/spring/boot/jdbc/SentryDsProxyAutoConfiguration {
 	public fun <init> ()V
 	public fun sentryQueryExecutionListener (Lio/sentry/IHub;)Lio/sentry/dsproxy/SentryQueryExecutionListener;
 }
 
-public class io/sentry/spring/boot/datasource/p6spy/SentryP6SpyAutoConfiguration {
+public class io/sentry/spring/boot/jdbc/SentryP6SpyAutoConfiguration {
 	public fun <init> ()V
 	public fun sentryJdbcEventListener (Lio/sentry/IHub;)Lio/sentry/p6spy/SentryJdbcEventListener;
 }
