+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/jdbc/SentryP6SpyAutoConfigurationTest.kt
@@ -1,4 +1,4 @@
-package io.sentry.spring.boot.datasource.p6spy
+package io.sentry.spring.boot.jdbc
 
 import io.sentry.p6spy.SentryJdbcEventListener
 import io.sentry.spring.boot.SentryAutoConfiguration
